# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Skoll API
  version: 1.0.0
  description: OpenAPI document for Skoll core HTTP endpoints.
servers:
  - url: __API_BASE_PREFIX__
paths:
...
  /v1/plugins/pharma_oa/api/employees:
    get:
      summary: List Pharma OA employees
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, on_leave, left]
        - name: offset
          in: query
          schema: {type: integer, minimum: 0, default: 0}
        - name: cursor
          in: query
          description: Opaque cursor returned by the previous page. When present, it takes precedence over offset.
          schema: {type: string}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 200, default: 50}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PharmaMasterListAPIResponse'}
        '403':
          description: Forbidden
    post:
      summary: Create Pharma OA employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, departmentId, positionId]
              properties:
                code:
                  type: string
                name:
                  type: string
                departmentId:
                  type: string
                positionId:
                  type: string
                phone:
                  type: string
                email:
                  type: string
                certificates:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Created
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/employees/{id}:
    put:
      summary: Update Pharma OA employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/employees/{id}/leave:
    post:
      summary: Mark Pharma OA employee left
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
  /v1/plugins/pharma_oa/api/employees/qualification-reminders:
    get:
      summary: List Pharma OA employee qualification reminders
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: OK
        '403':
          description: Forbidden
...
  /v1/plugins/pharma_oa/api/inventory-alerts:
    get:
      summary: List Pharma OA inventory alerts
      parameters:
        - {name: activeOnly, in: query, schema: {type: boolean, default: false}}
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryAlertListResponse'}
        '400': {description: Invalid activeOnly filter}
  /v1/plugins/pharma_oa/api/inventory-alert-jobs:
    get:
      summary: List Pharma OA inventory alert scan jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryAlertJobListResponse'}
    post:
      summary: Run a near-expiry, low-stock, and over-stock alert scan
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/InventoryAlertPolicyRequest'}
      responses:
        '201':
          description: Scan completed and actionable reminders emitted
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryAlertJobResponse'}
        '400': {description: Invalid thresholds, recipient, or scan failure}
  /v1/plugins/pharma_oa/api/inventory-alert-jobs/{id}/retry:
    post:
      summary: Retry a failed Pharma OA inventory alert scan job
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: Retry completed
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InventoryAlertJobResponse'}
        '400': {description: Job not found, not failed, or retry failed}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Skoll JWT access token. Aggregated plugin API operations require this scheme and declare x-skoll-permission.
  schemas:
...
    PharmaMasterListData:
      type: object
      required: [items, offset, limit, total, hasMore, nextCursor, sort]
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
        offset: {type: integer, minimum: 0}
        limit: {type: integer, minimum: 1, maximum: 200}
        total: {type: integer, format: int64, minimum: 0}
        hasMore: {type: boolean}
        nextCursor:
          type: string
          description: Empty on the final page; otherwise pass this value as the next request cursor.
        sort:
          type: string
          enum: ['code:asc']
    PharmaMasterListAPIResponse:
      allOf:
        - $ref: '#/components/schemas/MessageResponse'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/PharmaMasterListData'
...
    InventoryAlertPolicyRequest:
      type: object
      required: [nearExpiryDays, lowStockThreshold, overStockThreshold, recipientId]
      properties:
        nearExpiryDays: {type: integer, minimum: 1}
        lowStockThreshold: {type: integer, minimum: 0}
        overStockThreshold: {type: integer, minimum: 1}
        recipientId: {type: string}
    InventoryAlert:
      type: object
      required: [id, type, status, balanceId, productId, warehouseId, areaId, locationId, batchId, quantity, threshold, recipientId, notificationId, targetPath, firstSeenAt, lastSeenAt]
      properties:
        id: {type: string}
        type: {type: string, enum: [near_expiry, low_stock, over_stock]}
        status: {type: string, enum: [active, resolved]}
        balanceId: {type: string}
        productId: {type: string}
        warehouseId: {type: string}
        areaId: {type: string}
        locationId: {type: string}
        batchId: {type: string}
        quantity: {type: integer}
        threshold: {type: integer}
        expiresAt: {type: string, format: date-time}
        recipientId: {type: string}
        notificationId: {type: string}
        targetPath: {type: string}
        firstSeenAt: {type: string, format: date-time}
        lastSeenAt: {type: string, format: date-time}
        resolvedAt: {type: string, format: date-time}
    InventoryAlertJobLog:
      type: object
      required: [level, message, createdAt]
      properties:
        level: {type: string, enum: [info, error]}
        message: {type: string}
        createdAt: {type: string, format: date-time}
    InventoryAlertJob:
      type: object
      required: [id, status, policy, retryCount, matchedCount, createdCount, resolvedCount, logs, createdAt]
      properties:
        id: {type: string}
        status: {type: string, enum: [pending, running, succeeded, failed]}
        policy: {$ref: '#/components/schemas/InventoryAlertPolicyRequest'}
        retryCount: {type: integer, minimum: 0}
        matchedCount: {type: integer, minimum: 0}
        createdCount: {type: integer, minimum: 0}
        resolvedCount: {type: integer, minimum: 0}
        error: {type: string}
        logs: {type: array, items: {$ref: '#/components/schemas/InventoryAlertJobLog'}}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time}
        completedAt: {type: string, format: date-time}
    InventoryAlertListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: string}
        message: {type: string}
        data:
          type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/InventoryAlert'}}
    InventoryAlertJobResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: string}
        message: {type: string}
        data:
          type: object
          required: [item]
          properties:
            item: {$ref: '#/components/schemas/InventoryAlertJob'}
    InventoryAlertJobListResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: string}
        message: {type: string}
        data:
          type: object
          required: [items]
          properties:
            items: {type: array, items: {$ref: '#/components/schemas/InventoryAlertJob'}}
...
    MessageResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
...